requirements.txt
transformers @ git+https://github.com/huggingface/transformers
torch~=2.5.0
evalplus~=0.2.0
datasets>=2.18.0
trl~=0.8.6
peft~=0.10.0
python-dotenv~=1.0.1
scipy~=1.11.4
bitsandbytes~=0.43.1
einops~=0.7.0
evalplus~=0.2.1
datasets~=2.19.0
sentencepiece~=0.1.99
tree-sitter==0.21.3
accelerate @ git+https://github.com/huggingface/accelerate.git
astunparse
hydra-core
wandb~=0.17.1
tree_sitter_languages
fire~=0.6.0
tqdm~=4.66.2
numpy~=1.26.4
PyYAML~=6.0.1
pandas~=2.2.1
javalang~=0.13.0
setuptools~=70.0.0
matplotlib~=3.9.0